[WFLY-8382] Elytron capabilities interfaces into _private
