+++ elytron/src/main/java/org/wildfly/extension/elytron/capabilities/_private/CredentialSecurityFactory.java
@@ -15,22 +15,20 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.wildfly.extension.elytron.capabilities;
+package org.wildfly.extension.elytron.capabilities._private;
 
-import javax.naming.NamingException;
-import javax.naming.directory.DirContext;
-
-import org.wildfly.common.function.ExceptionSupplier;
+import org.wildfly.security.SecurityFactory;
+import org.wildfly.security.credential.Credential;
 
 /**
- * A {@link Supplier} for obtaining {@link DirContext} instances.
+ * A {@link SecurityFactory} that returns a {@link Credential}.
  *
  * @author <a href="mailto:darran.lofthouse@jboss.com">Darran Lofthouse</a>
  */
-public interface DirContextSupplier extends ExceptionSupplier<DirContext, NamingException> {
+public interface CredentialSecurityFactory extends SecurityFactory<Credential> {
 
-    static DirContextSupplier from(final ExceptionSupplier<DirContext, NamingException> supplier) {
-        return supplier::get;
+    static CredentialSecurityFactory from(final SecurityFactory<? extends Credential> function) {
+        return function::create;
     }
 
 }
